move code from loader.py to eic.py
