+++ python/sglang/srt/configs/model_config.py
@@ -599,18 +599,11 @@ class ModelConfig:
             # with statement to avoid closing the client.
             if is_eic(self.model_path):
                 from sglang.srt.connector.eic import pull_files_from_eic
-                model_name = self.model_path.removeprefix("eic://")
-                local_dir = "/tmp/" + "_".join([model_name, str(os.getpid())])
-                os.makedirs(local_dir, exist_ok=True)
-                logger.info(
-                    f"pull_files_from_eic model config start is_eic PID = {os.getpid()}, path {local_dir}")
-                assert os.path.isdir(local_dir), "dir not created!"
-                pull_files_from_eic(self.model_path, local_dir)
+                local_dir = pull_files_from_eic(self.model_path)
                 self.model_weights = self.model_path
                 self.model_path = local_dir
                 logger.info(
                     f"pull_files_from_eic model config end is_eic PID = {os.getpid()}, path {local_dir}")
-
             else:
                 client = create_remote_connector(self.model_path)
                 client.pull_files(allow_pattern=["*config.json"])
