+++ python/sglang/srt/connector/eic.py
@@ -1,7 +1,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import logging
-from typing import Generator, List, Optional, Tuple
+from typing import Generator, List, Optional, Tuple, Dict, Any
 
 import torch
 import ctypes
@@ -12,6 +12,9 @@ import logging
 import numpy as np
 import json
 import signal
+import math
+import ctypes
+import threading
 import eic
 import boto3
 from botocore.client import Config
@@ -96,21 +99,23 @@ def _get_config_file_key(model_path: str, filename: str) -> str:
     return os.path.join(model_path, filename)
 
 
-def pull_files_from_eic(model_path: str, local_dir: str):
-    logger.info(
-        f"pull_files_from_eic... local_dir = {local_dir} PID = {os.getpid()}")
+def pull_files_from_eic(model_path: str):
+    model_name = model_path.removeprefix("eic://")
+    local_dir = "/tmp/" + "_".join([model_name, str(os.getpid())])
+    os.makedirs(local_dir, exist_ok=True)
 
     pull_config_file = os.path.join(local_dir, EICConnector.PULL_CONFIG_FILE)
     if os.path.exists(pull_config_file):
         logger.info(
             f"config file exists, no need to pull again, local_dir = {local_dir} PID = {os.getpid()}")
-        return
+        return local_dir
 
     import multiprocessing
     p = multiprocessing.Process(
         target=do_pull_files_from_eic, args=(model_path, local_dir))
     p.start()
     p.join()
+    return local_dir
 
 
 def do_pull_files_from_eic(model_path: str, local_dir: str):
@@ -123,6 +128,13 @@ def do_pull_files_from_eic(model_path: str, local_dir: str):
 class EICConnector(BaseKVConnector):
     PULL_CONFIG_FILE = ".eic_model_config"
 
+    MODEL_BATCH_SET_SIZE = 20 * 1024 * 1024
+    MODEL_DATA_SPLIT_SIZE = 20 * 1024 * 1024
+    MODEL_META_APPROXIMATE_SIZE = 10 * 1024 * 1024
+    MODEL_META_PREFIX_MAX_LENGTH = 16
+    MODEL_META_PREFIX = "._meta"
+    ENABLE_ZERO_COPY = True
+    ENABLE_THREADING = False
     def __init__(self, url: str):
         super().__init__(url)
         for sig in (signal.SIGINT, signal.SIGTERM):
@@ -194,6 +206,11 @@ class EICConnector(BaseKVConnector):
 
         self.object_storage_client = s3_client
         logger.info(f"object client init success")
+        self.loaded_kv: dict[str, Any] = {}
+        self.perf_statistics: dict[str, Any] = {}
+        self.thread_num = 4
+        self.zero_copy_ptrs = []
+        self.split_tensors = []
 
     @classmethod
     def instance(cls, *args, **kwargs):
@@ -475,6 +492,410 @@ class EICConnector(BaseKVConnector):
             val = self.get(key)
             key = key.removeprefix(f"{self.model_name}/keys/rank_{rank}/")
             yield key, val
+
+    def _write_huge_kv(
+        self,
+        model_name: str,
+        key: str,
+        tensor: torch.Tensor,
+        data_split_size: int,
+    ) -> None:
+        # split tensor
+        tensor_size = tensor.nelement() * tensor.element_size()
+        split_num = math.ceil(tensor_size / data_split_size)
+        for i in range(split_num):
+            start = i * data_split_size
+            end = min((i + 1) * data_split_size, tensor_size)
+            split_tensor = tensor.view(
+                tensor.nelement()).view(torch.int8)[start: end]
+            split_key = os.path.join(model_name, key, str(i))
+            self.mset([split_key], [split_tensor])
+
+    def save_model(self, model_prefix, model_name, local_model_path, rank, state_dict):
+        start_time = time.perf_counter()
+        # root: ( k = model_prefix , v = { kv_num, meta_number, data_split_size, meta_prefix_name, kv_num_per_meta[] } )
+        root_json = {
+            "kv_num": 0,
+            "meta_num": 0,
+            "data_split_size": self.MODEL_DATA_SPLIT_SIZE,
+            "meta_prefix_name": self.MODEL_META_PREFIX,
+            "kv_num_per_meta": [],
+        }
+
+        # meta:( k = model_prefix/meta_prefix_name/{index} , v = { kv_num, key[], tensor_meta[], value_length[] })
+        meta_size = 0
+        meta_json = {
+            "kv_num": 0,
+            "key": [],
+            "tensor_meta": [],
+            "value_length": [],
+        }
+        # data: ( k = key, v = tensor )
+        # set tensors
+        keys: List[str] = []
+        tensors: List[torch.Tensor] = []
+        batch_size = 0
+        for key, tensor in state_dict.items():
+            should_write_data = False
+            # update meta
+            root_json["kv_num"] += 1
+            meta_json["kv_num"] += 1
+            tensor_size = tensor.nelement() * tensor.element_size()
+            meta_json["value_length"].append(tensor_size)
+            tensor_meta = generate_tensor_metadata(tensor)
+            meta_json["tensor_meta"].append(tensor_meta)
+            meta_json["key"].append(key)
+            # logger.info(f"model_prefix {model_prefix} save key {key} tensor_size {tensor_size} tensor_meta {tensor_meta}")
+            meta_size += len(key)
+            if tensor_size > self.MODEL_DATA_SPLIT_SIZE:
+                should_write_data = True
+                self._write_huge_kv(
+                    model_prefix,
+                    key,
+                    tensor,
+                    self.MODEL_DATA_SPLIT_SIZE,
+                )
+            else:
+                # append to write batch
+                keys.append(os.path.join(model_prefix, key))
+                tensors.append(tensor)
+                batch_size = batch_size + tensor_size
+                if batch_size >= self.MODEL_BATCH_SET_SIZE:
+                    should_write_data = True
+                    self.mset(keys, tensors)
+                    keys.clear()
+                    tensors.clear()
+                    batch_size = 0
+            # write meta if needed
+            if should_write_data and meta_size >= self.MODEL_META_APPROXIMATE_SIZE:
+                meta_name = os.path.join(
+                    model_prefix, self.MODEL_META_PREFIX, str(root_json["meta_num"]))
+                meta_bytes = json.dumps(meta_json).encode()
+                meta_buffer = torch.from_numpy(np.frombuffer(
+                    meta_bytes, dtype=np.int8).copy())
+                self.mset([meta_name], [meta_buffer])
+                # update root
+                root_json["kv_num_per_meta"].append(meta_json["kv_num"])
+                root_json["meta_num"] += 1
+                # reset counters
+                meta_json["kv_num"] = 0
+                meta_json["value_length"].clear()
+                meta_json["tensor_meta"].clear()
+                meta_json["key"].clear()
+                meta_size = 0
+
+        # write last data
+        if batch_size > 0:
+            self.mset(keys, tensors)
+        # write last meta
+        if meta_json["kv_num"] > 0:
+            meta_name = os.path.join(
+                model_prefix, self.MODEL_META_PREFIX, str(root_json["meta_num"]))
+            meta_bytes = json.dumps(meta_json).encode()
+            meta_buffer = torch.from_numpy(np.frombuffer(
+                meta_bytes, dtype=np.int8).copy())
+            self.mset([meta_name], [meta_buffer])
+            # update root
+            root_json["kv_num_per_meta"].append(meta_json["kv_num"])
+            root_json["meta_num"] += 1
+        # write root
+        logger.info(
+            f"write root {model_prefix}, content: {root_json}"
+        )
+        root_bytes = json.dumps(root_json).encode()
+        root_buffer = torch.from_numpy(np.frombuffer(
+            root_bytes, dtype=np.int8).copy())
+        self.mset([model_prefix], [root_buffer])
+
+        # upload files
+        file_list: List[str] = []
+        for root, _, files in os.walk(local_model_path):
+            for file_name in files:
+                # ignore hidden files
+                if file_name.startswith("."):
+                    continue
+                if os.path.splitext(file_name)[1] in (".jpg"):
+                    continue
+                if os.path.splitext(file_name)[1] not in (".bin", ".pt", ".safetensors"):
+                    file_list.append(os.path.join(root, file_name))
+        self.upload_model_config(model_name, file_list)
+        end_time = time.perf_counter()
+        logger.info(
+            f"save model on rank{rank} to eic cost {end_time - start_time}s")
+        # Preventing follower nodes from upload failures caused by the leader node exiting prematurely
+        # in multi node scenarios
+        time.sleep(30)
+
+    def _load_tensor(self, state_dict, key: str, tensor: torch.Tensor):
+
+        if key not in state_dict:
+            logger.warning(f"Key '{key}' not found in state_dict, replace key.")
+            if "self_attn.attn_mha." in key:
+                true_key = key.replace("self_attn.attn_mha.", "self_attn.")
+                key = true_key
+
+        param_data = state_dict[key].data
+        param_shape = state_dict[key].shape
+        # If loading with LoRA enabled, additional padding may
+        # be added to certain parameters. We only load into a
+        # narrowed view of the parameter data.
+        for dim, size in enumerate(tensor.shape):
+            if size < param_shape[dim]:
+                param_data = param_data.narrow(
+                    dim, 0, size)
+        if tensor.shape != param_shape:
+            logger.warning(
+                "loading tensor of shape %s into "
+                "parameter '%s' of shape %s",
+                tensor.shape,
+                key,
+                param_shape,
+            )
+        param_data.copy_(tensor)
+        state_dict.pop(key)
+
+    def _read_huge_kv(
+        self,
+        state_dict: Dict[str, torch.Tensor],
+        dict_key: str,
+        key: str,
+        value_length: int,
+        tensor_meta: Dict[str, any],
+        thread_index,
+    ) -> torch.Tensor:
+        start_time = time.perf_counter()
+        # read tensor
+        param_data = state_dict[dict_key].data
+        param_shape = state_dict[dict_key].shape
+        # If loading with LoRA enabled, additional padding may
+        # be added to certain parameters. We only load into a
+        # narrowed view of the parameter data.
+        for dim, size in enumerate(tensor_meta["shape"]):
+            if size < param_shape[dim]:
+                param_data = param_data.narrow(
+                    dim, 0, size)
+        if torch.Size(tensor_meta["shape"]) != param_shape:
+            logger.warning(
+                "loading tensor of shape %s into "
+                "parameter '%s' of shape %s",
+                tensor_meta["shape"],
+                key,
+                param_shape,
+            )
+        split_num = math.ceil(value_length / self.data_split_size)
+        for i in range(split_num):
+            split_start_time = time.perf_counter()
+            split_key = os.path.join(key, str(i))
+            start = i * self.data_split_size
+            end = min((i + 1) * self.data_split_size, value_length)
+            self.mget([split_key], [self.split_tensors[thread_index][0: end - start]])
+            split_end_time = time.perf_counter()
+            self.perf_statistics["read_huge_kv_split_s"] = self.perf_statistics.get(
+                "read_huge_kv_split_s", 0) + split_end_time - split_start_time
+            self.perf_statistics["read_huge_kv_split_ops"] = self.perf_statistics.get(
+                "read_huge_kv_split_ops", 0) + 1
+            param_data.view(torch.int8).view(value_length)[
+            start:end].copy_(self.split_tensors[thread_index][0: end - start])
+            load_end_time = time.perf_counter()
+            self.perf_statistics["read_huge_kv_copy_split_s"] = self.perf_statistics.get(
+                "read_huge_kv_copy_split_s", 0) + load_end_time - split_end_time
+        end_time = time.perf_counter()
+        self.perf_statistics["read_huge_kv_s"] = self.perf_statistics.get(
+            "read_huge_kv_s", 0) + end_time - start_time
+        self.perf_statistics["read_huge_kv_ops"] = self.perf_statistics.get(
+            "read_huge_kv_ops", 0) + 1
+        state_dict.pop(dict_key)
+        return
+
+    def _read_kv(
+        self,
+        keys: List[str],
+        tensor_metas: List[Dict[str, any]],
+    ) -> List[torch.Tensor]:
+        start_time = time.perf_counter()
+        tensors: List[torch.Tensor] = []
+        for i, meta in enumerate(tensor_metas):
+            tensors.append(get_tensor_from_metadata(meta))
+        self.mget(keys, tensors)
+        end_time = time.perf_counter()
+        self.perf_statistics["read_multi_kv_s"] = self.perf_statistics.get(
+            "read_multi_kv_s", 0) + end_time - start_time
+        self.perf_statistics["read_multi_kv_ops"] = self.perf_statistics.get(
+            "read_multi_kv_ops", 0) + 1
+        return tensors
+
+    def _read_meta(
+        self,
+        model_name: str,
+    ) -> None:
+        start_time = time.perf_counter()
+        # read meta and load to dict loader_kv
+        for i in range(self.meta_num):
+            meta_name = os.path.join(
+                model_name, self.meta_prefix, str(i))
+            tensors = self.mget([meta_name])
+            if len(tensors) == 0:
+                raise ValueError(f"Meta {meta_name} not found!")
+            meta_bytes = tensors[0].numpy().tobytes()
+            meta_json = json.loads(meta_bytes.decode())
+            meta_kv_num = meta_json["kv_num"]
+            if meta_kv_num != self.kv_per_meta[i]:
+                raise ValueError(
+                    f"Meta {meta_name} kv num {meta_kv_num} "
+                    f"not equal to root kv_per_meta {self.kv_per_meta[i]}")
+            for j in range(meta_kv_num):
+                key = meta_json["key"][j]
+                value_length = meta_json["value_length"][j]
+                tensor_meta = meta_json["tensor_meta"][j]
+                self.loaded_kv[key] = (value_length, tensor_meta)
+        end_time = time.perf_counter()
+        self.perf_statistics["read_meta_latency_s"] = end_time - start_time
+
+    def _read_root(
+        self,
+        root_name: str,
+    ) -> None:
+        start_time = time.perf_counter()
+        tensors = self.mget([root_name])
+        if len(tensors) == 0:
+            raise ValueError(f"Root {root_name} not found!")
+        root_bytes = tensors[0].numpy().tobytes()
+        root_json = json.loads(root_bytes.decode())
+        logger.info(
+            f"read root {root_name}, content: {root_json}"
+        )
+        self.kv_num = root_json["kv_num"]
+        self.meta_num = root_json["meta_num"]
+        self.data_split_size = root_json["data_split_size"]
+        self.meta_prefix = root_json["meta_prefix_name"]
+        self.kv_per_meta = root_json["kv_num_per_meta"]
+        end_time = time.perf_counter()
+        self.perf_statistics["read_root_latency_s"] = end_time - start_time
+        if self.ENABLE_ZERO_COPY:
+            if not self.ENABLE_THREADING:
+                self.thread_num = 1
+            for i in range(self.thread_num):
+                zero_copy_ptr = self.allocate_zero_copy_buffer(
+                    self.data_split_size)
+                ubyte_ptr = ctypes.cast(
+                    zero_copy_ptr, ctypes.POINTER(ctypes.c_ubyte))
+                byte_array = (
+                    ctypes.c_ubyte * self.data_split_size).from_address(ctypes.addressof(ubyte_ptr.contents))
+                data_bytes = memoryview(byte_array)
+                split_tensor = torch.frombuffer(data_bytes, dtype=torch.int8)
+                self.zero_copy_ptrs.append(zero_copy_ptr)
+                self.split_tensors.append(split_tensor)
+        else:
+            if not self.ENABLE_THREADING:
+                self.thread_num = 1
+            for i in range(self.thread_num):
+                split_tensor = torch.empty(
+                    self.data_split_size, dtype=torch.int8)
+                self.split_tensors.append(split_tensor)
+
+    def _load_eic_worker(self, model_name, state_dict, sub_items: list[tuple[str, Any]], thread_index):
+        # iterate all keys
+        batch_size = 0
+        eic_keys = []
+        state_dict_keys = []
+        tensor_metas = []
+        for key, meta in sub_items:
+            # logger.info(f"model_name {model_name} load key {key} meta {meta}")
+            value_length = meta[0]
+            tensor_meta = meta[1]
+            if value_length > self.data_split_size:
+                # read and copy tensor to state_dict
+                self._read_huge_kv(state_dict, key, os.path.join(
+                    model_name, key), value_length, tensor_meta, thread_index)
+            else:
+                eic_keys.append(os.path.join(model_name, key))
+                state_dict_keys.append(key)
+                batch_size = batch_size + value_length
+                tensor_metas.append(tensor_meta)
+                if batch_size > self.MODEL_DATA_SPLIT_SIZE:
+                    tensors = self._read_kv(eic_keys, tensor_metas)
+                    for i, tensor in enumerate(tensors):
+                        self._load_tensor(
+                            state_dict, state_dict_keys[i], tensor)
+                    # reset
+                    batch_size = 0
+                    eic_keys.clear()
+                    state_dict_keys.clear()
+                    tensor_metas.clear()
+        if batch_size > 0:
+            tensors = self._read_kv(eic_keys, tensor_metas)
+            for i, tensor in enumerate(tensors):
+                self._load_tensor(state_dict, state_dict_keys[i], tensor)
+
+    def load_model(self, model_prefix, state_dict, rank):
+        start_time = time.perf_counter()
+        # read root
+        self._read_root(model_prefix)
+        # read meta
+        self._read_meta(model_prefix)
+        # iterate all keys
+        batch_size = 0
+        eic_keys = []
+        state_dict_keys = []
+        tensor_metas = []
+        if self.ENABLE_THREADING:
+            items = list(self.loaded_kv.items())
+            chunk_size = (len(items) + self.thread_num - 1) // self.thread_num
+            threads = []
+            for i in range(self.thread_num):
+                start = i * chunk_size
+                end = min(start + chunk_size, len(items))
+                sub_items = items[start:end]
+                thread = threading.Thread(target=self._load_eic_worker,
+                                          args=(model_prefix, state_dict, sub_items, i))
+                threads.append(thread)
+                thread.start()
+            for thread in threads:
+                thread.join()
+        else:
+            for key, meta in self.loaded_kv.items():
+                # logger.info(f"model_prefix {model_prefix} load key {key} meta {meta}")
+                value_length = meta[0]
+                tensor_meta = meta[1]
+                if value_length > self.data_split_size:
+                    # read and copy tensor to state_dict
+                    self._read_huge_kv(state_dict, key, os.path.join(
+                        model_prefix, key), value_length, tensor_meta, 0)
+                else:
+                    eic_keys.append(os.path.join(model_prefix, key))
+                    state_dict_keys.append(key)
+                    batch_size = batch_size + value_length
+                    tensor_metas.append(tensor_meta)
+                    if batch_size > self.MODEL_DATA_SPLIT_SIZE:
+                        tensors = self._read_kv(eic_keys, tensor_metas)
+                        for i, tensor in enumerate(tensors):
+                            self._load_tensor(state_dict, state_dict_keys[i], tensor)
+                        # reset
+                        batch_size = 0
+                        eic_keys.clear()
+                        state_dict_keys.clear()
+                        tensor_metas.clear()
+            if batch_size > 0:
+                tensors = self._read_kv(eic_keys, tensor_metas)
+                for i, tensor in enumerate(tensors):
+                    self._load_tensor(state_dict, state_dict_keys[i], tensor)
+
+        if state_dict:
+            raise ValueError(
+                f"Missing keys {tuple(state_dict)} in loaded state!")
+
+        if self.ENABLE_ZERO_COPY:
+            for i in range(self.thread_num):
+                self.release_zero_copy_buffer(
+                    self.zero_copy_ptrs[i], self.data_split_size)
+            self.split_tensors = []
+            self.zero_copy_ptrs = []
+        else:
+            self.split_tensor = None
+        end_time = time.perf_counter()
+        self.perf_statistics["load_model_s"] = end_time - start_time
+        logger.info(
+            f"load model on rank{rank} from eic, perf_stat: {self.perf_statistics}")
 
     def close(self) -> None:
         if self.connection:
